ptx: updated code generation for map, generate
The 'map' operation always treats the array as its underlying representation, whereas 'generate' will produce nested loops to make use of multidimensional grids/thread blocks.

Note that we generate two versions, for 32-bit and 64-bit loop counters, and at execute time launch the appropriate version depending on the extent of the output array. For example, a two-dimensional `zipWith (+)` requires 13 registers when using 32-bit counters, but a whopping 36 when 64-bit counters are used. I'm not sure why the register counts are so inflated...

Note that this still launches the kernel as a one-dimensional index space.